RtGam: Estimating Rt with Generalized Additive Models

An opinionated implementation of a generalized additive modeling (GAMs) approach to estimate epidemic dynamics. This flexible semi-parametric method allows fitting to arbitrary incidence timeseries, with a natural Bayesian interpretation. Underlying inference methods are implemented via the mgcv package, which is optimized to produce quick, robust model fits and has well-developed diagnostic methods. These methods are employed here with a focus on _real-time_ use cases. The package is meant for drop-in use in an emerging outbreak, complementing MCMC-based inference methods.

See also

Author

Maintainer: Zachary Susswein utb2@cdc.gov (ORCID)

Authors:

Katelyn Gostic uep6@cdc.gov (ORCID)

Sam Abbott azw1@cdc.gov (ORCID)