Parse input vectors into a format for {mgcv}

Parse input vectors into a format for {mgcv}

Usage

dataset_creator(cases, reference_date, group, backend)

Arguments

cases

A vector of non-negative incident case counts occurring on an associated reference_date. Missing values (NAs) are not allowed.

reference_date

The associated date on which the count of incident cases occurred. Missing dates are not allowed and dates can only occur once.

group

The grouping variable for the case/reference-date pair. Not yet implemented and a value other than NULL will throw an error.

backend

One of gam or bam; defaults to gam. In general, models should be fit with mgcv::gam(). If mgcv::gam() is too slow, mgcv::bam() converges more quickly but introduces some additional numerical error. Note that the bam backend uses the discrete = TRUE option for an additional speedup. See mgcv::bam() for more information.

Value

A dataframe for mgcv