Validate user-provided inputs for type correctness

Does not validate sensibility of inputs nor does it do any comparisons between the inputs to check lengths

Usage

validate(cases, reference_date, group, k, m, call = rlang::caller_env())

Arguments

cases

A vector of non-negative incident case counts occurring on an associated reference_date. Missing values (NAs) are not allowed.

reference_date

The associated date on which the count of incident cases occurred. Missing dates are not allowed and dates can only occur once.

group

The grouping variable for the case/reference-date pair. Not yet implemented and a value other than NULL will throw an error.

k

An integer, the total dimension of all the smoothing basis functions. Defaults to smooth_dim_heuristic(length(cases)), which picks a reasonable estimate based on the number of provided data points. This total dimension is partitioned between the different smooths in the model. In the case of diagnostic issues in a fitted RtGam model, increasing the value of k above this default and refitting the model is a good first step. See the smooth_dim_heuristic() documentation for more information.

m

An integer, the dimension of the penalty basis for the global smooth trend. If m is greater than 1, the smooth's wiggliness can change over time. An increase in this value above the default should be done carefully. See penalty_dim_heuristic() for more information on m and when to consider changing the default.

call

Caller environment to be passed through the stack